move gdsfactory.utils to gdsfactory.technology
